requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
apscheduler==3.10.4
celery==5.4.0
redis==5.2.1
requests==2.32.3
beautifulsoup4==4.12.3
yfinance==0.2.54
pandas==2.1.4
numpy==1.26.4
python-dotenv==1.0.1
